Conflicts:
	src/plugins/netstoremanager/managertab.cpp

+++ src/plugins/sb2/desaturateeffect.h
@@ -1,6 +1,6 @@
 /**********************************************************************
  * LeechCraft - modular cross-platform feature rich internet client.
- * Copyright (C) 2006-2013  Georg Rudoy
+ * Copyright (C) 2006-2014  Georg Rudoy
  *
  * Boost Software License - Version 1.0 - August 17th, 2003
  *
@@ -41,7 +41,7 @@ namespace SB2
 
 		Q_PROPERTY (qreal strength READ GetStrength WRITE SetStrength NOTIFY strengthChanged)
 
-		qreal Strength_;
+		float Strength_;
 	public:
 		DesaturateEffect (QObject* = 0);
